addition

+++ 0x00-python_variable_annotations/0-add.py
@@ -2,6 +2,6 @@
 """ returning a basic addition"""
 
 
-def add(a: int, b: int) -> int:
+def add(a: float, b: float) -> float:
     """returning a basic addition"""
     return a + b
